Move HEC raw receiver to merge with HEC receiver, simplify HEC receiver and drop path config setting
