--- receiver/splunkhecrawreceiver/receiver_test.go
@@ -1,469 +1,0 @@
-// Copyright 2020, OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package splunkhecrawreceiver
-
-import (
-	"bytes"
-	"compress/gzip"
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/config/confighttp"
-	"go.opentelemetry.io/collector/config/configtls"
-	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/model/pdata"
-	"go.uber.org/zap"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/testutil"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/splunk"
-)
-
-func Test_splunkhecreceiver_NewLogsReceiver(t *testing.T) {
-	defaultConfig := createDefaultConfig().(*Config)
-	emptyEndpointConfig := createDefaultConfig().(*Config)
-	emptyEndpointConfig.Endpoint = ""
-	type args struct {
-		config       Config
-		logsConsumer consumer.Logs
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr error
-	}{
-		{
-			name: "nil_nextConsumer",
-			args: args{
-				config: *defaultConfig,
-			},
-			wantErr: errNilNextLogsConsumer,
-		},
-		{
-			name: "empty_endpoint",
-			args: args{
-				config:       *emptyEndpointConfig,
-				logsConsumer: new(consumertest.LogsSink),
-			},
-			wantErr: errEmptyEndpoint,
-		},
-		{
-			name: "default_endpoint",
-			args: args{
-				config:       *defaultConfig,
-				logsConsumer: consumertest.NewNop(),
-			},
-		},
-		{
-			name: "happy_path",
-			args: args{
-				config: Config{
-					HTTPServerSettings: confighttp.HTTPServerSettings{
-						Endpoint: "localhost:1234",
-					},
-				},
-				logsConsumer: consumertest.NewNop(),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := newLogsReceiver(zap.NewNop(), tt.args.config, tt.args.logsConsumer)
-			assert.Equal(t, tt.wantErr, err)
-			if err == nil {
-				assert.NotNil(t, got)
-			} else {
-				assert.Nil(t, got)
-			}
-		})
-	}
-}
-
-func Test_splunkhecReceiver_handleReq(t *testing.T) {
-	config := createDefaultConfig().(*Config)
-	config.Endpoint = "localhost:0" // Actually not creating the endpoint
-	config.initialize()
-
-	currentTime := float64(time.Now().UnixNano()) / 1e6
-	splunkMsg := buildSplunkHecMsg(currentTime, 3)
-
-	tests := []struct {
-		name           string
-		req            *http.Request
-		assertResponse func(t *testing.T, status int, body string)
-	}{
-		{
-			name: "incorrect_method",
-			req:  httptest.NewRequest("PUT", "http://localhost/foo", nil),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusBadRequest, status)
-				assert.Equal(t, responseInvalidMethod, body)
-			},
-		},
-		{
-			name: "incorrect_content_type",
-			req: func() *http.Request {
-				req := httptest.NewRequest("POST", "http://localhost/foo", nil)
-				req.Header.Set("Content-Type", "application/not-json")
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusOK, status)
-			},
-		},
-		{
-			name: "incorrect_content_encoding",
-			req: func() *http.Request {
-				req := httptest.NewRequest("POST", "http://localhost/foo", nil)
-				req.Header.Set("Content-Encoding", "superzipper")
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusUnsupportedMediaType, status)
-				assert.Equal(t, responseInvalidEncoding, body)
-			},
-		},
-		{
-			name: "empty_body",
-			req: func() *http.Request {
-				req := httptest.NewRequest("POST", "http://localhost/foo", bytes.NewReader(nil))
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusOK, status)
-			},
-		},
-
-		{
-			name: "two_logs",
-			req: func() *http.Request {
-				req := httptest.NewRequest("POST", "http://localhost/foo", strings.NewReader("foo\nbar"))
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusAccepted, status)
-			},
-		},
-		{
-			name: "msg_accepted",
-			req: func() *http.Request {
-				msgBytes, err := json.Marshal(splunkMsg)
-				require.NoError(t, err)
-				req := httptest.NewRequest("POST", "http://localhost/foo", bytes.NewReader(msgBytes))
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusAccepted, status)
-			},
-		},
-		{
-			name: "msg_accepted_gzipped",
-			req: func() *http.Request {
-				msgBytes, err := json.Marshal(splunkMsg)
-				require.NoError(t, err)
-
-				var buf bytes.Buffer
-				gzipWriter := gzip.NewWriter(&buf)
-				_, err = gzipWriter.Write(msgBytes)
-				require.NoError(t, err)
-				require.NoError(t, gzipWriter.Close())
-
-				req := httptest.NewRequest("POST", "http://localhost/foo", &buf)
-				req.Header.Set("Content-Encoding", "gzip")
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusAccepted, status)
-			},
-		},
-		{
-			name: "bad_gzipped_msg",
-			req: func() *http.Request {
-				msgBytes, err := json.Marshal(splunkMsg)
-				require.NoError(t, err)
-
-				req := httptest.NewRequest("POST", "http://localhost/foo", bytes.NewReader(msgBytes))
-				req.Header.Set("Content-Encoding", "gzip")
-				return req
-			}(),
-			assertResponse: func(t *testing.T, status int, body string) {
-				assert.Equal(t, http.StatusBadRequest, status)
-				assert.Equal(t, responseErrGzipReader, body)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			sink := new(consumertest.LogsSink)
-			rcv, err := newLogsReceiver(zap.NewNop(), *config, sink)
-			assert.NoError(t, err)
-
-			r := rcv.(*splunkReceiver)
-			assert.NoError(t, r.Start(context.Background(), componenttest.NewNopHost()))
-			defer r.Shutdown(context.Background())
-			w := httptest.NewRecorder()
-			r.handleReq(w, tt.req)
-
-			resp := w.Result()
-			respBytes, err := ioutil.ReadAll(resp.Body)
-			assert.NoError(t, err)
-
-			var bodyStr string
-			if len(respBytes) > 0 {
-				assert.NoError(t, json.Unmarshal(respBytes, &bodyStr))
-			} else {
-				bodyStr = ""
-			}
-
-			tt.assertResponse(t, resp.StatusCode, bodyStr)
-		})
-	}
-}
-
-func Test_consumer_err(t *testing.T) {
-	currentTime := float64(time.Now().UnixNano()) / 1e6
-	splunkMsg := buildSplunkHecMsg(currentTime, 3)
-	config := createDefaultConfig().(*Config)
-	config.Endpoint = "localhost:0" // Actually not creating the endpoint
-	config.initialize()
-	rcv, err := newLogsReceiver(zap.NewNop(), *config, consumertest.NewErr(errors.New("bad consumer")))
-	assert.NoError(t, err)
-
-	r := rcv.(*splunkReceiver)
-	r.Start(context.Background(), componenttest.NewNopHost())
-	defer r.Shutdown(context.Background())
-	w := httptest.NewRecorder()
-	msgBytes, err := json.Marshal(splunkMsg)
-	require.NoError(t, err)
-	req := httptest.NewRequest("POST", "http://localhost", bytes.NewReader(msgBytes))
-	r.handleReq(w, req)
-
-	resp := w.Result()
-	respBytes, err := ioutil.ReadAll(resp.Body)
-	assert.NoError(t, err)
-
-	var bodyStr string
-	assert.NoError(t, json.Unmarshal(respBytes, &bodyStr))
-
-	assert.Equal(t, http.StatusInternalServerError, resp.StatusCode)
-	assert.Equal(t, "Internal Server Error", bodyStr)
-}
-
-func Test_splunkhecReceiver_TLS(t *testing.T) {
-	addr := testutil.GetAvailableLocalAddress(t)
-	cfg := createDefaultConfig().(*Config)
-	cfg.Endpoint = addr
-	cfg.TLSSetting = &configtls.TLSServerSetting{
-		TLSSetting: configtls.TLSSetting{
-			CertFile: "./testdata/server.crt",
-			KeyFile:  "./testdata/server.key",
-		},
-	}
-	cfg.initialize()
-	sink := new(consumertest.LogsSink)
-	r, err := newLogsReceiver(zap.NewNop(), *cfg, sink)
-	require.NoError(t, err)
-	defer r.Shutdown(context.Background())
-
-	mh := newAssertNoErrorHost(t)
-	require.NoError(t, r.Start(context.Background(), mh), "should not have failed to start log reception")
-
-	// If there are errors reported through host.ReportFatalError() this will retrieve it.
-	<-time.After(500 * time.Millisecond)
-	t.Log("Event Reception Started")
-
-	logs := pdata.NewLogs()
-	rl := logs.ResourceLogs().AppendEmpty()
-	ill := rl.InstrumentationLibraryLogs().AppendEmpty()
-	lr := ill.Logs().AppendEmpty()
-
-	lr.Body().SetStringVal("foo")
-	want := logs
-
-	t.Log("Sending Splunk HEC data Request")
-
-	url := fmt.Sprintf("https://%s", addr)
-	req, err := http.NewRequest("POST", url, strings.NewReader("foo"))
-	require.NoErrorf(t, err, "should have no errors with new request: %v", err)
-
-	tlscs := configtls.TLSClientSetting{
-		TLSSetting: configtls.TLSSetting{
-			CAFile:   "./testdata/ca.crt",
-			CertFile: "./testdata/client.crt",
-			KeyFile:  "./testdata/client.key",
-		},
-		ServerName: "localhost",
-	}
-	tls, errTLS := tlscs.LoadTLSConfig()
-	assert.NoError(t, errTLS)
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: tls,
-		},
-	}
-
-	resp, err := client.Do(req)
-	require.NoErrorf(t, err, "should not have failed when sending to splunk HEC receiver %v", err)
-	assert.Equal(t, http.StatusAccepted, resp.StatusCode)
-	t.Log("Splunk HEC Request Received")
-
-	got := sink.AllLogs()
-	require.Equal(t, 1, len(got))
-	assert.Equal(t, want, got[0])
-}
-
-func Test_splunkhecReceiver_AccessTokenPassthrough(t *testing.T) {
-	tests := []struct {
-		name        string
-		passthrough bool
-		token       pdata.AttributeValue
-	}{
-		{
-			name:        "No token provided and passthrough false",
-			passthrough: false,
-			token:       pdata.NewAttributeValueNull(),
-		},
-		{
-			name:        "No token provided and passthrough true",
-			passthrough: true,
-			token:       pdata.NewAttributeValueNull(),
-		},
-		{
-			name:        "token provided and passthrough false",
-			passthrough: false,
-			token:       pdata.NewAttributeValueString("myToken"),
-		},
-		{
-			name:        "token provided and passthrough true",
-			passthrough: true,
-			token:       pdata.NewAttributeValueString("myToken"),
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			config := createDefaultConfig().(*Config)
-			config.Endpoint = "localhost:0"
-			config.AccessTokenPassthrough = tt.passthrough
-			config.initialize()
-
-			sink := new(consumertest.LogsSink)
-			rcv, err := newLogsReceiver(zap.NewNop(), *config, sink)
-			assert.NoError(t, err)
-
-			currentTime := float64(time.Now().UnixNano()) / 1e6
-			splunkhecMsg := buildSplunkHecMsg(currentTime, 3)
-			msgBytes, _ := json.Marshal(splunkhecMsg)
-			req := httptest.NewRequest("POST", "http://localhost", bytes.NewReader(msgBytes))
-			if tt.token.Type() != pdata.AttributeValueTypeNull {
-				req.Header.Set("Splunk", tt.token.StringVal())
-			}
-
-			r := rcv.(*splunkReceiver)
-			r.Start(context.Background(), componenttest.NewNopHost())
-			defer r.Shutdown(context.Background())
-			w := httptest.NewRecorder()
-			r.handleReq(w, req)
-
-			resp := w.Result()
-			respBytes, err := ioutil.ReadAll(resp.Body)
-			assert.NoError(t, err)
-
-			var bodyStr string
-
-			if len(respBytes) > 0 {
-				assert.NoError(t, json.Unmarshal(respBytes, &bodyStr))
-			} else {
-				bodyStr = ""
-			}
-
-			assert.Equal(t, http.StatusAccepted, resp.StatusCode)
-
-			got := sink.AllLogs()
-
-			resource := got[0].ResourceLogs().At(0).Resource()
-			tokenLabel, exists := resource.Attributes().Get("com.splunk.hec.access_token")
-
-			if tt.passthrough {
-				if tt.token.Type() == pdata.AttributeValueTypeNull {
-					assert.False(t, exists)
-				} else {
-					assert.Equal(t, tt.token.StringVal(), tokenLabel.StringVal())
-				}
-			} else {
-				assert.Empty(t, tokenLabel)
-			}
-		})
-	}
-}
-
-func buildSplunkHecMsg(time float64, dimensions uint) *splunk.Event {
-	ev := &splunk.Event{
-		Time:       &time,
-		Event:      "foo",
-		Fields:     map[string]interface{}{},
-		Index:      "myindex",
-		SourceType: "custom:sourcetype",
-	}
-	for dim := uint(0); dim < dimensions; dim++ {
-		ev.Fields[fmt.Sprintf("k%d", dim)] = fmt.Sprintf("v%d", dim)
-	}
-
-	return ev
-}
-
-type badReqBody struct{}
-
-var _ io.ReadCloser = (*badReqBody)(nil)
-
-func (b badReqBody) Read(p []byte) (n int, err error) {
-	return 0, errors.New("badReqBody: can't read it")
-}
-
-func (b badReqBody) Close() error {
-	return nil
-}
-
-// assertNoErrorHost implements a component.Host that asserts that there were no errors.
-type assertNoErrorHost struct {
-	component.Host
-	*testing.T
-}
-
-// newAssertNoErrorHost returns a new instance of assertNoErrorHost.
-func newAssertNoErrorHost(t *testing.T) component.Host {
-	return &assertNoErrorHost{
-		Host: componenttest.NewNopHost(),
-		T:    t,
-	}
-}
-
-func (aneh *assertNoErrorHost) ReportFatalError(err error) {
-	assert.NoError(aneh, err)
-}
