--- receiver/splunkhecrawreceiver/testdata/ca.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDNjCCAh4CCQCnfoiEG7LBSDANBgkqhkiG9w0BAQsFADBdMQswCQYDVQQGEwJB
-VTESMBAGA1UECAwJQXVzdHJhbGlhMQ8wDQYDVQQHDAZTeWRuZXkxEjAQBgNVBAoM
-CU15T3JnTmFtZTEVMBMGA1UEAwwMTXlDb21tb25OYW1lMB4XDTIxMDUwNjAzMTYz
-MVoXDTMxMDUwNDAzMTYzMVowXTELMAkGA1UEBhMCQVUxEjAQBgNVBAgMCUF1c3Ry
-YWxpYTEPMA0GA1UEBwwGU3lkbmV5MRIwEAYDVQQKDAlNeU9yZ05hbWUxFTATBgNV
-BAMMDE15Q29tbW9uTmFtZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
-AK5ZXfEgUVgALVG3xmFL3VILH94GsKOTnCXkcsd/cYaPt6iPRBhkMSMFaHDmIjFG
-SwFXs6EffPWlNthhDDI33ZfESiunYK0kYGpE721aD+kB1jme8mpWFP3TAuRa29JO
-Q5Fv0Rk/QdmY5NK5Ym6P2wMNA1EOXXCnvDuIG3j5fH9VsfDV4k0I3PcvZPD5DRKH
-U08SBjWRi2cs1ozKdQprQMSOA+iqeRrubq5DJykiHnZpye6Kv6BED7vX7ffp9g3u
-wfptSlYqydrCCMV4Qgln0cP2RmDeKJsReCosnEqB3jskv3wCotCkKIHIQqOPmzmN
-/cV8gBQR59DO5X2I6w0hhk0CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEABEhgETCJ
-dXMAIxTVwZPz52Kn+PtdkL7Ewjy91vLxwvzApMRQSdEIrrClyffay/26jnADX/uT
-H+UnsphBvQelB14ler8GsKGUek59IXr6NxsPeME/hBVM7vmfgx2lwL8uDx3yPg30
-35LqJuinOyr+kuzumPGG0DqxPs38xRu39IrSKW/PrHRwlTKEvoLU2ugYqdwz9TV+
-Ec0OEEmdrBUV8V0G8f03UxTJ/WFZolTRcHvmIaSzJw7PkMycjwbcKMYU6xObhae+
-6lw+huUdQ6yYAIfMBz50zlx6uSkwEyU7DDddtF6J/ALUPoZDvu+8cPnEOFDHm8hT
-JnPBg0kB3JDSNw==
------END CERTIFICATE-----
