--- receiver/splunkhecrawreceiver/testdata/client.crt
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDazCCAlOgAwIBAgIJAJ04Tjxr61SIMA0GCSqGSIb3DQEBCwUAMF0xCzAJBgNV
-BAYTAkFVMRIwEAYDVQQIDAlBdXN0cmFsaWExDzANBgNVBAcMBlN5ZG5leTESMBAG
-A1UECgwJTXlPcmdOYW1lMRUwEwYDVQQDDAxNeUNvbW1vbk5hbWUwHhcNMjEwNTA2
-MDMxNjMxWhcNMzEwNTA0MDMxNjMxWjBdMQswCQYDVQQGEwJBVTESMBAGA1UECAwJ
-QXVzdHJhbGlhMQ8wDQYDVQQHDAZTeWRuZXkxEjAQBgNVBAoMCU15T3JnTmFtZTEV
-MBMGA1UEAwwMTXlDb21tb25OYW1lMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEApP2GNuoJh0DSwHN/pJLF7xeoCnLm4xUlD20gKxS3DLiXGOc42vmrBAqn
-M25dxto3FByzHSjaMpzaWdM7VDDvgzbNcdWvOy0juQfq16TiVGTouRycrl7WpbPr
-5qB0BW6W831UZMa2EsL35vcEKgernToZCRkKh014PIPXB6jMCsMCLsAU6jNLrtJx
-ot2X7cXqP9EFuQqqxIFn4y2u/xJ68sTYA/Vt0uDYOmDEcSs8rUPjBDDiIpcq79Qg
-4Z3kgeHB1uRWXyarBK20LsGJn9ukeWP9XhfZS3nLDkmirygiKHJbx09DOI8mcr8a
-GRoKFjwSau3llFLsNWXgitp7o1y6jwIDAQABoy4wLDAUBgNVHREEDTALgglsb2Nh
-bGhvc3QwFAYDVR0RBA0wC4IJbG9jYWxob3N0MA0GCSqGSIb3DQEBCwUAA4IBAQCT
-YnS7J4zuDSrplUPG+o1LGoxIFmqyecTrzjsMHRxuiss9R8Zp2Nkw6lUC2JrjXlWY
-8dxsab45quq0Af80UKSdE9ZEUSbdGqxLixTJ2nTLNIo1eeHqhXZpLBHzE9Bd5YMB
-waSfwknaX2W/b625VQOHD1otmZUBfwJj4DlTgj+vhEH2TITogxCfgF/kTia4h58P
-6EjDb5tKG3HhCvOTIVAauiPBLJrW/MdyH7liljt78PqumXGd2WEnMsLEYvUpgfX3
-VEeSSaMTzAViI2j9xx08yXoY18VjX/lZrX5fdoQmrJupp351oXrV1snsAetwD/MZ
-FbkboKvJHkVvZAEUGJoO
------END CERTIFICATE-----
