--- receiver/splunkhecrawreceiver/testdata/client.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEogIBAAKCAQEApP2GNuoJh0DSwHN/pJLF7xeoCnLm4xUlD20gKxS3DLiXGOc4
-2vmrBAqnM25dxto3FByzHSjaMpzaWdM7VDDvgzbNcdWvOy0juQfq16TiVGTouRyc
-rl7WpbPr5qB0BW6W831UZMa2EsL35vcEKgernToZCRkKh014PIPXB6jMCsMCLsAU
-6jNLrtJxot2X7cXqP9EFuQqqxIFn4y2u/xJ68sTYA/Vt0uDYOmDEcSs8rUPjBDDi
-Ipcq79Qg4Z3kgeHB1uRWXyarBK20LsGJn9ukeWP9XhfZS3nLDkmirygiKHJbx09D
-OI8mcr8aGRoKFjwSau3llFLsNWXgitp7o1y6jwIDAQABAoIBABonimAoJ4yGTSJu
-7LwU/EWTlXLalhlHX9GbNe8+YGom1OUAunWUAHUqjyjLBRihvPt2NTqZIEXHsaFt
-p8kH3OALDvRNsHKMDE6nn6XBoJNKT84QaBIIPLSV2f58TwgRaOJ+aqNzEKfWul+l
-psZW4vCQpilYC+A4+qd1KuAu9PF7iSK6yShOG8NgGfC6Ms1X6lMp1O0G9ceiqbUF
-Jx3CSo6OVcgJHtg/fQYIi6HXKqcFBkniCE3Fw1Cy72F0bWXH/daTLdtDYoXzuaCc
-k8clJKBD/hKJsoJf81u3jOrw9g/YS00PaAcpd5wSuy4EeR0Ld9VqZ75TizsViI3x
-KuoBHaECgYEA1uj/zYiQPoh6l21pluCOh38JPu67CnDMR/bz8gXjWtK0+yDs+9ce
-rtM+ITfYqrRbaRvFNj2FB+JqR/izAPLWuEUUewCXqEj2yB7E8hBS0B3mX5SCbZfY
-pjL89Wy+DGOyHfE+HrnnTiAyob0IJA25wSjsvQQPgQeIIcjiCdttGHECgYEAxIkl
-uv8wEpu2ttPxigty+hlQMHsdwbniey2HYUjjncaXboiHAbU59jWwxxekg/irboac
-W72xMf+VXLUTn/ZOSxPFQB4Kl8GloloBKm44oJNcvICTpTaJxAxv0xJqdF8x3BMp
-Cucl16TRb3IVUgxv9mhdiQqZ3yz11CjOXEPIgv8CgYBsxzrTOdBpFoOWqhbL+5Wp
-F7ACXqvluPptYp5PXpOZoRViF2RchKJPMzU7cx0EKo4qQZgswC7ZsszA6S0RydjN
-JIodSEt7LpNa79gX/hMsrDWzI2RxPG64kQ5g203Gz5+UcEtVGhvt+3YlT50TQq9H
-CgPqGGgQk1NgZmGjQ6GQwQKBgHWNTpsFjZxjtKqwJFWcOX4r69HDoe3AoB1rAwGG
-VlaFdQfw80drPSrtvDmpA4bIEsm+DAG3eSfdplnYGM/hw7yWK7IJBe71LXvo96+I
-U6h/28UH5se30jkH8069Qh0u3ySx9SeHUBQmifptsyDrPYCVtfSE+boKK1NlwOBg
-zqrDAoGAGivjugMj/8nnpvEJjnMxmkiwN5NDgjcwA+GWFiHXpcAn/inw9VByk1HN
-HiJvrcvDWjGVw8ZuyQV3XD7MWwMYAy9PZnjc0yy4KknJc/xqooxUywkg+5dASxfU
-vfuVw6CO+YLZbkrTruSba7SPxoPBozZJCDWml0YTYiE7SmjxqEE=
------END RSA PRIVATE KEY-----
