--- receiver/splunkhecrawreceiver/testdata/config.yaml
@@ -1,24 +1,0 @@
-receivers:
-  splunk_hec_raw:
-  splunk_hec_raw/allsettings:
-    # endpoint specifies the network interface and port which will receive
-    # Splunk metrics.
-    endpoint: localhost:8088
-    access_token_passthrough: true
-  splunk_hec_raw/tls:
-    tls_settings:
-      cert_file: /test.crt
-      key_file: /test.key
-
-processors:
-  nop:
-
-exporters:
-  nop:
-
-service:
-  pipelines:
-    metrics:
-      receivers: [splunk_hec_raw, splunk_hec_raw/allsettings]
-      processors: [nop]
-      exporters: [nop]
