--- receiver/splunkhecrawreceiver/testdata/server.crt
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDazCCAlOgAwIBAgIJAJ04Tjxr61SHMA0GCSqGSIb3DQEBCwUAMF0xCzAJBgNV
-BAYTAkFVMRIwEAYDVQQIDAlBdXN0cmFsaWExDzANBgNVBAcMBlN5ZG5leTESMBAG
-A1UECgwJTXlPcmdOYW1lMRUwEwYDVQQDDAxNeUNvbW1vbk5hbWUwHhcNMjEwNTA2
-MDMxNjMxWhcNMzEwNTA0MDMxNjMxWjBdMQswCQYDVQQGEwJBVTESMBAGA1UECAwJ
-QXVzdHJhbGlhMQ8wDQYDVQQHDAZTeWRuZXkxEjAQBgNVBAoMCU15T3JnTmFtZTEV
-MBMGA1UEAwwMTXlDb21tb25OYW1lMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEA7tZiXRp2PDEb/PL5+jLEJ8rjmdxXZyOgwlcF5gLXeCCmWgpQWnYBiRnR
-qT3N+/5ngWzw40yRsire/LeiOpyC9zi0MyxMxJYJBKmYpK6oY+9EA95aOqdta4yP
-H6qabi2FduLjPMHmkbff0D8iJraDFnabLBYJrg6YV6ARo/dG7mLDZbL7LetMon51
-/d5+89j80YWVx5SmiRgEc43upljvBNDYC3YDirlWmhBbilXVLj9c+KrVMi4HuXSb
-DxyssEJiB0P+xCPdvKqYJaHIGHelXWtKFgYva/BYs+HnbgU7UkRxQdNnvQ1cCbmM
-cZycOiyV3dgzTMiZrP8IBREXnXuAEQIDAQABoy4wLDAUBgNVHREEDTALgglsb2Nh
-bGhvc3QwFAYDVR0RBA0wC4IJbG9jYWxob3N0MA0GCSqGSIb3DQEBCwUAA4IBAQBY
-1ncvHEIz4hsmJriti/ydvEEnhSsLyD2Ejwf+X27iqmGWPA/8w1O+VD1oXlrW+rAT
-vnmZZBcQ9acMlt6Ctp8kGx0dcJN2k41n8I8LUTjvuYATZHMg9j6wtGO4SgtVlspZ
-Pz05N6O6ltogA/hiZPtCXhp86UiJ9TvE/EcV6pGaDcfIJP/71DD7QtytHQE/aHap
-FTHvH8gIqBSR23RlCu+hxiPPQ79EY/cSYbrRsGFmnXL3/+XXSMqc7EKXl3uWEIXW
-xXE+oOukOWbzDbbadjkgeL063fQr48mSN9R+QonQrCfh2y9EiXlmi8t3SI2KdjZ/
-a6nS3+2OZvxSzzoo6nfx
------END CERTIFICATE-----
