--- receiver/splunkhecrawreceiver/testdata/server.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEA7tZiXRp2PDEb/PL5+jLEJ8rjmdxXZyOgwlcF5gLXeCCmWgpQ
-WnYBiRnRqT3N+/5ngWzw40yRsire/LeiOpyC9zi0MyxMxJYJBKmYpK6oY+9EA95a
-Oqdta4yPH6qabi2FduLjPMHmkbff0D8iJraDFnabLBYJrg6YV6ARo/dG7mLDZbL7
-LetMon51/d5+89j80YWVx5SmiRgEc43upljvBNDYC3YDirlWmhBbilXVLj9c+KrV
-Mi4HuXSbDxyssEJiB0P+xCPdvKqYJaHIGHelXWtKFgYva/BYs+HnbgU7UkRxQdNn
-vQ1cCbmMcZycOiyV3dgzTMiZrP8IBREXnXuAEQIDAQABAoIBAQDIihIFgtaMK+AI
-QCxNPJsfnmgFJNHutCejhlZZj0FEiIWlTMuzaSpY5+j3ZgGy+SYHsmWBNSWQqBab
-s8Tv2cLNujr0TNoYprHjzFNy04RohNpjNSwl27cwetEiEamhoW9TJ6cNuPHUoMO4
-o4oVX2IqVmRsXAS8tVyNFsY5GsPgHEX8zdR7oI+mgro9UrwqvMkX5ktC85IsmJk4
-otSDldg9kmLUheRbefX9Y7dvbjfZfJ7cSxvSL2Ygr0lTyQ8NsMNhxgq1nNNX9beh
-AwDBkTBK2/UlIuVAUxJCwSiKVfmbzTWJEqkTciOepZCnfHFIFgGdsEVgp7yo3Mw2
-RFRtP6ABAoGBAP2J3iytC1Z5+EvD4xJiQkRt33YKQfzO1CmOWiK7rVKp0KIuLsCV
-YgAE3vIJf6A9vWQA3nx8jcF5lfiOToMcSYgiKPLcgnc1OTwx2fhc/KuEsTOmn6cV
-zyXsT3RiXM3DTLmXEcFXSW4AsJ0vBtwRBafghYeCdUdteepWouVMl3+RAoGBAPEn
-+p0SaKCUx0q8VONJEPFr+EkLim2VXqSlK2YPWxK1pnVv0Y6wfywx9qRZFLK8WGYo
-3Dgenrp6peMw+0G4JfQj2Pu8Rjtj83ki+oMP0xkgqsEI4oxqVlUye1c1klb1kB3W
-ObjUNjEPaYRd5b7y37AGInqVqFth2KU7Rvm4NLiBAoGBANtl2qS+efkzQvINaQoX
-jVC54426ArBKKr5rOQSlc6781FYDsOFmY4+wp1y7a32275UCS4VeIy0xkd2tB7Pd
-XqQ3K0g+GYgevpCgcwbBMMUb5lvT7IQiE0e13mvQUo6cWmCfofW1Gw3d+veKIl19
-PW+wktxYwDQ3QDDRF9VlGhSBAoGAOZKP1Dv22o9qXL9FjNV+xARAzD+DFtJYO/pq
-t/zmgzr2uIavgt9EHSr4dRgHRvyMJiePG57DykVJoUmqUnBCL1z3inHSWo9sKvlc
-0TlP7mnMGjWi0F76qXBZ3eIeQ8BwXY8tTKDtJx1E1aMNHF8jZASy6o0cGmTrsdbL
-BQcjHwECgYAtKU9BQAgtkjIjcAypDSORqR+qz29lMm8E1sKRfEM0H/deHJmb5tEa
-1XeMffcAAOzj9zpanb3KgjxPzx+taYTdJn/H+8e4JaNn7Xbkpk8Fi7kp+qHdUD9O
-/5oAbTg8sgBDqu7JU3TWH9g3EUYyFxGWGRqqJXRS3oFdFlDArXoVxg==
------END RSA PRIVATE KEY-----
