+++ receiver/splunkhecreceiver/factory.go
@@ -53,7 +53,6 @@ func createDefaultConfig() config.Receiver {
 			Endpoint: defaultEndpoint,
 		},
 		AccessTokenPassthroughConfig: splunk.AccessTokenPassthroughConfig{},
-		Path:                         "",
 	}
 }
