+++ receiver/splunkhecreceiver/receiver.go
@@ -15,22 +15,22 @@
 package splunkhecreceiver
 
 import (
+	"bufio"
 	"compress/gzip"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io/ioutil"
 	"net"
 	"net/http"
 	"sync"
 	"time"
 
 	"github.com/gorilla/mux"
-	"go.opencensus.io/trace"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
 	"go.opentelemetry.io/collector/obsreport"
 	"go.uber.org/zap"
 
@@ -59,9 +59,10 @@ var (
 	errNilNextMetricsConsumer = errors.New("nil metricsConsumer")
 	errNilNextLogsConsumer    = errors.New("nil logsConsumer")
 	errEmptyEndpoint          = errors.New("empty endpoint")
+	errInvalidMethod          = errors.New("invalid http method")
+	errInvalidEncoding        = errors.New("invalid encoding")
 
 	okRespBody                = initJSONResponse(responseOK)
-	notFoundRespBody          = initJSONResponse(responseNotFound)
 	invalidMethodRespBody     = initJSONResponse(responseInvalidMethod)
 	invalidEncodingRespBody   = initJSONResponse(responseInvalidEncoding)
 	errGzipReaderRespBody     = initJSONResponse(responseErrGzipReader)
@@ -170,6 +171,9 @@ func (r *splunkReceiver) Start(_ context.Context, host component.Host) error {
 
 	mx := mux.NewRouter()
 	mx.NewRoute().HandlerFunc(r.handleReq)
+	if r.metricsConsumer == nil {
+		mx.NewRoute().Path(r.config.RawPath).HandlerFunc(r.handleRawReq)
+	}
 
 	r.server = r.config.HTTPServerSettings.ToServer(mx, r.settings)
 
@@ -193,6 +197,70 @@ func (r *splunkReceiver) Shutdown(context.Context) error {
 	return r.server.Close()
 }
 
+func (r *splunkReceiver) handleRawReq(resp http.ResponseWriter, req *http.Request) {
+	ctx := req.Context()
+	ctx = r.obsrecv.StartLogsOp(ctx)
+
+	if req.Method != http.MethodPost {
+		r.failRequest(ctx, resp, http.StatusBadRequest, invalidMethodRespBody, 0, errInvalidMethod)
+		return
+	}
+
+	encoding := req.Header.Get(httpContentEncodingHeader)
+	if encoding != "" && encoding != gzipEncoding {
+		r.failRequest(ctx, resp, http.StatusUnsupportedMediaType, invalidEncodingRespBody, 0, errInvalidEncoding)
+		return
+	}
+
+	if req.ContentLength == 0 {
+		r.obsrecv.EndLogsOp(ctx, typeStr, 0, nil)
+		return
+	}
+
+	bodyReader := req.Body
+	if encoding == gzipEncoding {
+		reader := r.gzipReaderPool.Get().(*gzip.Reader)
+		err := reader.Reset(bodyReader)
+
+		if err != nil {
+			r.failRequest(ctx, resp, http.StatusBadRequest, errGzipReaderRespBody, 0, err)
+			_, _ = ioutil.ReadAll(req.Body)
+			_ = req.Body.Close()
+			return
+		}
+		bodyReader = reader
+		defer r.gzipReaderPool.Put(reader)
+	}
+
+	sc := bufio.NewScanner(bodyReader)
+
+	ld := pdata.NewLogs()
+	rl := ld.ResourceLogs().AppendEmpty()
+	if r.config.AccessTokenPassthrough {
+		accessToken := req.Header.Get(splunk.HECTokenHeader)
+		if accessToken != "" {
+			rl.Resource().Attributes().InsertString(splunk.HecTokenLabel, accessToken)
+		}
+	}
+	ill := rl.InstrumentationLibraryLogs().AppendEmpty()
+
+	for sc.Scan() {
+		logRecord := ill.Logs().AppendEmpty()
+		logLine := sc.Text()
+		logRecord.Body().SetStringVal(logLine)
+	}
+	decodeErr := r.logsConsumer.ConsumeLogs(ctx, ld)
+
+	_ = bodyReader.Close()
+
+	if decodeErr != nil {
+		r.failRequest(ctx, resp, http.StatusInternalServerError, errInternalServerError, ill.Logs().Len(), decodeErr)
+	} else {
+		resp.WriteHeader(http.StatusAccepted)
+		r.obsrecv.EndLogsOp(ctx, typeStr, ill.Logs().Len(), nil)
+	}
+}
+
 func (r *splunkReceiver) handleReq(resp http.ResponseWriter, req *http.Request) {
 	ctx := req.Context()
 	if r.logsConsumer == nil {
@@ -200,20 +268,15 @@ func (r *splunkReceiver) handleReq(resp http.ResponseWriter, req *http.Request) 
 	} else {
 		ctx = r.obsrecv.StartLogsOp(ctx)
 	}
-	reqPath := req.URL.Path
-	if !r.config.pathGlob.Match(reqPath) {
-		r.failRequest(ctx, resp, http.StatusNotFound, notFoundRespBody, nil)
-		return
-	}
 
 	if req.Method != http.MethodPost {
-		r.failRequest(ctx, resp, http.StatusBadRequest, invalidMethodRespBody, nil)
+		r.failRequest(ctx, resp, http.StatusBadRequest, invalidMethodRespBody, 0, errInvalidMethod)
 		return
 	}
 
 	encoding := req.Header.Get(httpContentEncodingHeader)
 	if encoding != "" && encoding != gzipEncoding {
-		r.failRequest(ctx, resp, http.StatusUnsupportedMediaType, invalidEncodingRespBody, nil)
+		r.failRequest(ctx, resp, http.StatusUnsupportedMediaType, invalidEncodingRespBody, 0, errInvalidEncoding)
 		return
 	}
 
@@ -222,7 +285,7 @@ func (r *splunkReceiver) handleReq(resp http.ResponseWriter, req *http.Request) 
 		reader := r.gzipReaderPool.Get().(*gzip.Reader)
 		err := reader.Reset(bodyReader)
 		if err != nil {
-			r.failRequest(ctx, resp, http.StatusBadRequest, errGzipReaderRespBody, err)
+			r.failRequest(ctx, resp, http.StatusBadRequest, errGzipReaderRespBody, 0, err)
 			return
 		}
 		bodyReader = reader
@@ -242,16 +305,16 @@ func (r *splunkReceiver) handleReq(resp http.ResponseWriter, req *http.Request) 
 		var msg splunk.Event
 		err := dec.Decode(&msg)
 		if err != nil {
-			r.failRequest(ctx, resp, http.StatusBadRequest, errUnmarshalBodyRespBody, err)
+			r.failRequest(ctx, resp, http.StatusBadRequest, errUnmarshalBodyRespBody, len(events), err)
 			return
 		}
 		if msg.IsMetric() {
 			if r.metricsConsumer == nil {
-				r.failRequest(ctx, resp, http.StatusBadRequest, errUnsupportedMetricEvent, err)
+				r.failRequest(ctx, resp, http.StatusBadRequest, errUnsupportedMetricEvent, len(events), err)
 				return
 			}
 		} else if r.logsConsumer == nil {
-			r.failRequest(ctx, resp, http.StatusBadRequest, errUnsupportedLogEvent, err)
+			r.failRequest(ctx, resp, http.StatusBadRequest, errUnsupportedLogEvent, len(events), err)
 			return
 		}
 
@@ -264,25 +327,34 @@ func (r *splunkReceiver) handleReq(resp http.ResponseWriter, req *http.Request) 
 	}
 }
 
-func (r *splunkReceiver) createResourceCustomizer(req *http.Request) func(pdata.Resource) {
+func (r *splunkReceiver) createResourceCustomizer() func(*http.Request, pdata.Resource) {
 	if r.config.AccessTokenPassthrough {
-		if accessToken := req.Header.Get(splunk.HECTokenHeader); accessToken != "" {
-			return func(resource pdata.Resource) {
+		return func(req *http.Request, resource pdata.Resource) {
+			if accessToken := req.Header.Get(splunk.HECTokenHeader); accessToken != "" {
 				resource.Attributes().InsertString(splunk.HecTokenLabel, accessToken)
 			}
 		}
 	}
-	return func(resource pdata.Resource) {}
+	return func(req *http.Request, resource pdata.Resource) {}
 }
 
 func (r *splunkReceiver) consumeMetrics(ctx context.Context, events []*splunk.Event, resp http.ResponseWriter, req *http.Request) {
-	md, _ := splunkHecToMetricsData(r.settings.Logger, events, r.createResourceCustomizer(req), r.config)
+	var resourceCustomer func(resource pdata.Resource)
+	if r.config.AccessTokenPassthrough {
+		accessToken := req.Header.Get(splunk.HECTokenHeader)
+		if accessToken != "" {
+			resourceCustomer = func(resource pdata.Resource) {
+				resource.Attributes().InsertString(splunk.HecTokenLabel, accessToken)
+			}
+		}
+	}
+	md, _ := splunkHecToMetricsData(r.settings.Logger, events, resourceCustomer, r.config)
 
 	decodeErr := r.metricsConsumer.ConsumeMetrics(ctx, md)
 	r.obsrecv.EndMetricsOp(ctx, typeStr, len(events), decodeErr)
 
 	if decodeErr != nil {
-		r.failRequest(ctx, resp, http.StatusInternalServerError, errInternalServerError, decodeErr)
+		r.failRequest(ctx, resp, http.StatusInternalServerError, errInternalServerError, len(events), decodeErr)
 	} else {
 		resp.WriteHeader(http.StatusAccepted)
 		resp.Write(okRespBody)
@@ -290,16 +362,25 @@ func (r *splunkReceiver) consumeMetrics(ctx context.Context, events []*splunk.Ev
 }
 
 func (r *splunkReceiver) consumeLogs(ctx context.Context, events []*splunk.Event, resp http.ResponseWriter, req *http.Request) {
-	ld, err := splunkHecToLogData(r.settings.Logger, events, r.createResourceCustomizer(req), r.config)
+	var resourceCustomer func(resource pdata.Resource)
+	if r.config.AccessTokenPassthrough {
+		accessToken := req.Header.Get(splunk.HECTokenHeader)
+		if accessToken != "" {
+			resourceCustomer = func(resource pdata.Resource) {
+				resource.Attributes().InsertString(splunk.HecTokenLabel, accessToken)
+			}
+		}
+	}
+	ld, err := splunkHecToLogData(r.settings.Logger, events, resourceCustomer, r.config)
 	if err != nil {
-		r.failRequest(ctx, resp, http.StatusBadRequest, errUnmarshalBodyRespBody, err)
+		r.failRequest(ctx, resp, http.StatusBadRequest, errUnmarshalBodyRespBody, len(events), err)
 		return
 	}
 
 	decodeErr := r.logsConsumer.ConsumeLogs(ctx, ld)
-
+	r.obsrecv.EndLogsOp(ctx, typeStr, len(events), decodeErr)
 	if decodeErr != nil {
-		r.failRequest(ctx, resp, http.StatusInternalServerError, errInternalServerError, decodeErr)
+		r.failRequest(ctx, resp, http.StatusInternalServerError, errInternalServerError, len(events), decodeErr)
 	} else {
 		resp.WriteHeader(http.StatusAccepted)
 		resp.Write(okRespBody)
@@ -311,41 +392,34 @@ func (r *splunkReceiver) failRequest(
 	resp http.ResponseWriter,
 	httpStatusCode int,
 	jsonResponse []byte,
+	numRecordsReceived int,
 	err error,
 ) {
 	resp.WriteHeader(httpStatusCode)
 	if len(jsonResponse) > 0 {
 		// The response needs to be written as a JSON string.
+		resp.Header().Add("Content-Type", "application/json")
 		_, writeErr := resp.Write(jsonResponse)
 		if writeErr != nil {
 			r.settings.Logger.Warn("Error writing HTTP response message", zap.Error(writeErr))
 		}
 	}
 
-	msg := string(jsonResponse)
+	if r.metricsConsumer == nil {
+		r.obsrecv.EndLogsOp(ctx, typeStr, numRecordsReceived, err)
+	} else {
+		r.obsrecv.EndMetricsOp(ctx, typeStr, numRecordsReceived, err)
+	}
 
-	reqSpan := trace.FromContext(ctx)
-	reqSpan.AddAttributes(
-		trace.Int64Attribute(conventions.AttributeHTTPStatusCode, int64(httpStatusCode)),
-		trace.StringAttribute("http.status_text", msg))
-	traceStatus := trace.Status{
-		Code: trace.StatusCodeInvalidArgument,
+	if r.settings.Logger.Core().Enabled(zap.DebugLevel) {
+		msg := string(jsonResponse)
+		r.settings.Logger.Debug(
+			"Splunk HEC receiver request failed",
+			zap.Int("http_status_code", httpStatusCode),
+			zap.String("msg", msg),
+			zap.Error(err), // It handles nil error
+		)
 	}
-	if httpStatusCode == http.StatusInternalServerError {
-		traceStatus.Code = trace.StatusCodeInternal
-	}
-	if err != nil {
-		traceStatus.Message = err.Error()
-	}
-	reqSpan.SetStatus(traceStatus)
-	reqSpan.End()
-
-	r.settings.Logger.Debug(
-		"Splunk HEC receiver request failed",
-		zap.Int("http_status_code", httpStatusCode),
-		zap.String("msg", msg),
-		zap.Error(err), // It handles nil error
-	)
 }
 
 func initJSONResponse(s string) []byte {
