+++ receiver/splunkhecreceiver/testdata/config.yaml
@@ -5,7 +5,7 @@ receivers:
     # Splunk metrics.
     endpoint: localhost:8088
     access_token_passthrough: true
-    path: "/foo"
+    raw_path: "/foo"
     hec_metadata_to_otel_attrs:
       source: "file.name"
       sourcetype: "foobar"
